requirements.txt
onnx==1.10.2
numpy==1.21.0
pandas==1.4.1
pillow==9.0.1
django==3.2.12
waitress==2.1.1
gunicorn==20.1.0
requests==2.25.0
whitenoise==5.3.0
onnxruntime==1.9.0
django-heroku==0.3.1
opencv-python-headless
django-cors-headers==3.11.0